move error.py up one level
