+++ terminusdb_client/errors.py
